attempted build error fixes
