+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,101 +1,5 @@
-import NextAuth, { type NextAuthOptions } from "next-auth";
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import { prisma } from "@/lib/prisma";
-import GoogleProvider from "next-auth/providers/google";
-import CredentialsProvider from "next-auth/providers/credentials";
-import bcrypt from "bcrypt";
-import { getToken } from "next-auth/jwt";
-
-export const authOptions: NextAuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID || "",
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET || "",
-    }),
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "email" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          return null;
-        }
-
-        const user = await prisma.users.findUnique({
-          where: {
-            email: credentials.email,
-          },
-        });
-
-        if (!user) {
-          return null;
-        }
-
-        if (!user.password) {
-          throw new Error("Please sign in with Google.");
-        }
-
-        const passwordValid = await bcrypt.compare(
-          credentials.password,
-          user.password
-        );
-
-        if (!passwordValid) {
-          return null;
-        }
-
-        return {
-          id: user.id,
-          email: user.email,
-          name: user.name,
-          image: user.image,
-        };
-      },
-    }),
-  ],
-  session: {
-    strategy: "jwt",
-  },
-  pages: {
-    signIn: "/sign-in",
-    signOut: "/sign-out",
-    error: "/error",
-  },
-  callbacks: {
-    async session({ session, token }) {
-      if (token && session.user) {
-        const user = await prisma.users.findUnique({
-          where: { id: token.sub || "" }
-        });
-
-        if (user) {
-          session.user.role = user.role || "GUEST";
-          session.user.id = user.id;
-        }
-      }
-      return session;
-    },
-    async jwt({ token, user }) {
-      if (user) {
-        token.role = user.role || "GUEST";
-      }
-      return token;
-    },
-    async redirect({ url, baseUrl }) {
-      // PENDING users no longer redirected for demo purposes
-      
-      // Handle standard OAuth signups
-      if (url.includes('/sign-up') && url.includes('callbackUrl')) {
-        return url;
-      }
-      
-      return url.startsWith(baseUrl) ? url : baseUrl;
-    }
-  },
-};
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth"; // Import from the new central location
 
 const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
