+++ src/pages/api/check-in.ts
@@ -1,9 +1,10 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { prisma } from '@/lib/prisma';
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '@/app/api/auth/[...nextauth]/route';
+import { authOptions } from '@/lib/auth';
 import { Prisma } from '@prisma/client';
 import { mapSnakeToCamel, mapCamelToSnake } from '@/lib/map-utils';
+import { Shift, User } from '@/types';
 
 interface ResponseData {
   success: boolean;
